add admin acl condition

+++ app/Http/Controllers/DeviceController.php
@@ -9,22 +9,29 @@ use Illuminate\Support\Str;
 
 class DeviceController extends Controller
 {
+
     public function list(Request $request)
     {
-        $devices = Device::paginate(15);
+        $is_admin = auth()->user()->is_admin;
+        $user_id = auth()->user()->id;
+        $devices = $is_admin ? Device::paginate(15) : Device::where('user_id', $user_id)->paginate(15);
         return response()->json($devices);
     }
 
     public function view($id, Request $request) {
-        $device = Device::find($id);
+        $is_admin = auth()->user()->is_admin;
+        $user_id = auth()->user()->id;
+        $device = $is_admin ? Device::find($id) : Device::where('user_id', $user_id)->where('id', $id)->firstOrFail();
         return response()->json($device);
     }
 
     public function store(Request $request) {
+        $is_admin = auth()->user()->is_admin;
+        $user_id = auth()->user()->id;
         $id = $request->input('id', null);
         $id = $id != '' ? $id : null;
         if($id) {
-            $device = Device::find($id);
+            $device = $is_admin ? Device::find($id) : Device::where('user_id', $user_id)->where('id', $id)->firstOrFail();
         } else {
             $device = new Device();
             $device->password = Str::random(13);
@@ -36,7 +43,9 @@ class DeviceController extends Controller
     }
 
     public function delete($id, Request $request) {
-        $device = Device::where('id', $id)->delete();
+        $is_admin = auth()->user()->is_admin;
+        $user_id = auth()->user()->id;
+        $device = $is_admin ? Device::where('id', $id)->delete() : Device::where('user_id', $user_id)->where('id', $id)->delete();
         return response()->json(['message' => 'deleted']);
     }
 }
